requirements.txt
torch==2.1.0
sentence-transformers==2.2.2
ctransformers
transformers
numpy 
faiss-cpu
pandas
beautifulsoup4
flask